fix(composition): normalize network tick rate type

+++ Assets/VoxelEngine/Composition/NetworkingComposition.cs
@@ -421,7 +421,7 @@ namespace VoxelEngine.Composition
     /// <summary>Composition entry point for the concrete Net.Runtime implementation.</summary>
     public static class NetworkingComposition
     {
-        public static int AuthoritativeTickRateHz => AuthoritativeTickConfig.TickRateHz;
+        public static int AuthoritativeTickRateHz => (int)AuthoritativeTickConfig.TickRateHz;
 
         public static NetworkServerFacade CreateServer(uint serverSeed, int maxConnections = 64)
         {
